+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * 功能：注解 `@DbField` 新增 `name` 属性，可显式指定字段的参数名
   * 功能：注解 `@DbField` 新增 `cluster` 属性，可显式指定该字段是否是聚合字段
   * 功能：注解 `@SearchBean` 新增 `fields` 属性，用于指定额外的条件字段
+  * 优化：当 `@SearchBean.groupBy` 使用拼接参数时，条件字段的生成逻辑
 * Bean Searcher Boot Starter
   * GroupPairResolver 支持注入 Bean 的方式自定义
 * Bean Searcher Solon Plugin